Sem: CPlayer raw pointer purge
Idea of using raw pointers to pass CPlayer into another methods
was abolished completely as the program switched to using shared/weak
pointers everywhere. The reference to players opponent is now stored
as a weak_ptr as it is essentially a cyclical reference.

As the program needs to call the CDisplay::refresh_board with a pointer
to 'this', additional inheritance of CPlayer from
std::enable_shared_from_this was needed.