Sem: CDisplay::print_cards
Parts of refresh_board() method were moved into a new, separate method
that is able to print a collection (vector) of cards to a given position.
It only prints a maximum amount of cards (that can fit to screen) and it
returns the amount of unprinted cards (that is displayed to screen
later on).