Sem: CCardDeckable, card structure rework
Completely changed the inheritance from CCard as a middleman was inserted
in the process. Now, CCardWarcry and CCardTroop do not inherit straight
from CCard, but instead from this new middleman. This new inheritance
is now purely polymorphic, unlike before when it contained the problematic
class CCardGeneral. This seemingly minor change had consequences on
the whole codebase.